+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- feat: use remote config API ([#233](https://github.com/PostHog/posthog-android/pull/233))
+
 ## 3.12.0 - 2025-03-10
 
 - feat: support reuse of `anonymousId` between user changes ([#229](https://github.com/PostHog/posthog-android/pull/229))